+++ src/Events/Authorization/CloudEventXacmlMapper.cs
@@ -6,7 +6,6 @@ using System.Security.Claims;
 using Altinn.Authorization.ABAC.Xacml.JsonProfile;
 using Altinn.Common.PEP.Constants;
 using Altinn.Common.PEP.Helpers;
-using Altinn.Platform.Events.Helpers;
 using Altinn.Platform.Events.Models;
 
 using static Altinn.Authorization.ABAC.Constants.XacmlConstants;
